perf: simplified and fixed the ECM estimation
fixed and simplified the ECM estimation for both the adjpin and the mpin models.

(-) The optimal parameters have a limited number of decimals (5 for probabilities, and 3 for intensity rates). This improves the speed of the ECM estimation.

(+) The convergence condition is now the absolute difference between two consecutive loglikelihoods.

(+) Fixed errors in the calculation of the likelihood of the adjpin loglikelihood.

(+) fixed an error in the calculation of the uninformed trading rate on the sell side for the generation of the initial sets.