Fixed error in changing the number of cores in parallel computing
Earlier, the code automatically sets the number of parallel cores used for parallel computing to 2 whenever the number of cores in the machine is larger than 2. Now, the user can set freely the number of parallel cores used in computation as long as it does not exceed the number of cores available in the machine.