feat: Update initials_adjpin() algorithm to align with Ersan and Ghachem (2024)
Modified the implementation of the algorithm for generating the initial parameter sets in the initials_adjpin() function to align with the new algorithm described in Ersan and Ghachem (2024). Refer to the paper for more details.

Reference:  Ersan, Oguz and Ghachem, Montasser, A Methodological Approach to the Computational Problems in the Estimation of Adjusted Pin Model. Available at SSRN: https://ssrn.com/abstract=4359991 or http://dx.doi.org/10.2139/ssrn.4359991