feat: Update initials_adjpin_rnd() to align with Ersan and Ghachem (2024) algorithm
Updated the implementation of the algorithm for generating random initial parameter sets in initials_adjpin_rnd() to align with the method described in the paper by Ersan and Ghachem (2024). Refer to the paper for more details.

Reference:  Ersan, Oguz and Ghachem, Montasser, A Methodological Approach to the Computational Problems in the Estimation of Adjusted Pin Model. Available at SSRN: https://ssrn.com/abstract=4359991 or http://dx.doi.org/10.2139/ssrn.4359991